% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{accurCalc}
\alias{accurCalc}
\title{calculate accuracy from the predicted and observed values in
the prediction functions of dynamic Bayesian networks}
\usage{
accurCalc(df)
}
\description{
calculate accuracy from the predicted and observed values in
the prediction functions of dynamic Bayesian networks
}
\keyword{internal}
