% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{classficator}
\alias{classficator}
\title{Replace values with classes generated in the
 \code{\link{defClasses}} function}
\usage{
classficator(input, varclass, class_names)
}
\description{
Replace values with classes generated in the
 \code{\link{defClasses}} function
}
\keyword{internal}
