% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Dbn.R
\name{createDbn}
\alias{createDbn}
\title{Creates dynamic Bayesian networks for performance evaluation}
\usage{
createDbn(area)
}
\arguments{
\item{area}{A dataframe of continuous data, from a specific area
of the plantation to be used}
}
\value{
A dataframe with the dynamics networks metrics (RMSE, R-squared, MAE)
}
\description{
Using harvest data from all phenological phases of the cultivar,
from a specific area of the plantation, dynamic Bayesian networks
are generated (using the \code{ dbnR::learn_dbn_struc} functions with
\code{natPsoho and dmmhc} methods),
trained (using the \code{dbnR::fit_dbn_params} function)
and evaluated for performance (using \code{dbnR::forecast_ts} and
\code{ caret::defaultSummary}).
Two networks are created, one using \code{natPsoho} learning method and
other using \code{dmmhc} method.
The DBN are tested using \code{dbnR::forecast_ts} with \code{exact}
and \code{approx} methods then, from the return of these functions,
the RMSE, R-squared and MAE metrics are calculated and returned
}
\examples{
\donttest{
metricsDbn = createDbn(area_1)
}
}
