% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bn.R
\name{createNetworks}
\alias{createNetworks}
\title{Creates, from a dataframes, four Bayesian networks for performance evaluation}
\usage{
createNetworks(areaphase, blacklist, whitelist)
}
\arguments{
\item{areaphase}{the data.frame to be used}

\item{blacklist}{a dataframe of character string with two columns,
it is passed as a parameter to bnlearn learn functions in order to
avoid these arcs composing the final network}

\item{whitelist}{a dataframe of character string with two columns,
it is passed as a parameter to bnlearn learn functions in order to
guarantee these arcs composing the final network}
}
\value{
Network evaluation metrics,
as calculated in the \code{validateNetwork} function
}
\description{
Using harvest data from a phenological phase of the cultivar,
from a specific area of the plantation, Bayesian network are generated
(using the bnlearn::hc and bnlearn::mmhc functions),
trained (using the bnlearn::bn.fit function) and evaluated for performance.
Four networks are created, two from the pre-established topology and
two learned only from the presented data.
}
\keyword{internal}
