% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bn_test.R
\name{createNetworksTest}
\alias{createNetworksTest}
\title{Creates, from a dataframes, four Bayesian networks for performance evaluation}
\usage{
createNetworksTest(areaphase, areatype)
}
\arguments{
\item{areaphase}{the data.frame to be used}

\item{areatype}{integer to identify which type of area the dataframe belongs to}
}
\description{
From n crops data of a phenological phase in a expecific area in the cropfield,
four
A partir de dados de colheitas de uma fase fenológica da cultivar, de uma área
especícica da plantação, são geradas  (utilizando as funções  bnlearn::hc e
bnlearn::mmhc), treinadas (utilizando bnlearn::bn.fit) e
avaliadas quanto a performance (validaRede()) quatro redes bayesianas
}
