% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{defClasses}
\alias{defClasses}
\title{Defines classes for discretization of results from dynamic Bayesian networks}
\usage{
defClasses(n_classes, df_nome)
}
\description{
Defines classes for discretization of results from dynamic Bayesian networks
}
\keyword{internal}
