% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{f_minmax}
\alias{f_minmax}
\title{Allows normalization of generated data.}
\usage{
f_minmax(x)
}
\arguments{
\item{x}{the set of data to be normalized}
}
\value{
set of data normalized
}
\description{
Allows normalization of data using the $$(x−min)/(max−min)$$ equation.
}
\examples{
\donttest{
normalizerData = f_minmax(notNormalizerData)
}
}
\keyword{internal}
