% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{geraGraph}
\alias{geraGraph}
\title{Generate graphs of static bayesian networks metrics}
\usage{
geraGraph(areaN, file, subtitle)
}
\arguments{
\item{areaN}{dataframe with the data to be plotted on the chart}

\item{file}{string that indicates the path where the graph will be saved}

\item{subtitle}{string of the term that will be placed at top-left of the graph}
}
\description{
generate graphs of the accuracy of the four static Bayesian networks
generated in the \code{\link{testCreateNetworks}} and
\code{\link{createNetworks}} functions. The generated graphs will be
grouped by phenological phase.
}
\examples{
\donttest{
file = path_to_where_to_save_the_graph
area1bn <- testRunNetworks(areas_dis[[1]],1)
geraGraph(area1bn, file, "Area 1")
}
}
\keyword{internal}
