% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bn.R
\name{runNetworks}
\alias{runNetworks}
\title{Executes \code{\link{createNetworks}} function to an area}
\usage{
runNetworks(arealist, blacklist, whitelist)
}
\arguments{
\item{arealist}{list os dataframes of all phenological phase on
a specific area of the plantation}

\item{blacklist}{a dataframe of character string with two columns,
it is passed as a parameter to bnlearn learn functions in order to
avoid these arcs composing the final network}

\item{whitelist}{a dataframe of character string with two columns,
it is passed as a parameter to bnlearn learn functions in order to
guarantee these arcs composing the final network}
}
\value{
A dataframe organizing the network metrics generated in the
\code{\link{createNetworks}} function. The lines represent the network
performance learned in each phenological phase of the area (\code{arealist})
}
\description{
Executes the \code{\link{createNetworks}} function for all phenological
phases of an area. It also organizes the generated metrics in a dataframe.
}
\examples{
\donttest{
blacklist = data.frame(
from = c("X_1","X_1",
         "X_2", "X_2",
         "X_3", "X_3",
         "harvest",  "harvest",  "harvest" ),
to = c("X_2", "X_3", #from v1
       "X_1", "X_3", #from v2
       "X_1", "X_2", #from v3
       "X_1", "X_2", "X_3")) #from harvest
whitelist = data.frame(
from = c("X_1", "X_2", "X_3"),
to = c("harvest", "harvest", "harvest"))
arealist <- list(area1)
metricsArea1 <- createNetworks (arealist, blacklist, whitelist)
}
}
