% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_createBN}
\alias{test_createBN}
\title{TEST FUNCTION - tests generation of static Bayesian networks}
\usage{
test_createBN()
}
\description{
TEST FUNCTION - tests generation of static Bayesian networks
}
\keyword{internal}
