% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_createDBN}
\alias{test_createDBN}
\title{TEST FUNCTION - tests generation of static Dynamic Bayesian networks}
\usage{
test_createDBN()
}
\description{
TEST FUNCTION - tests generation of static Dynamic Bayesian networks
}
\keyword{internal}
