% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{testBuildSimulationData}
\alias{testBuildSimulationData}
\title{Initialize the values of all variables}
\usage{
testBuildSimulationData(nHarvests, nPhases, nAreas = NULL, ...)
}
\arguments{
\item{nHarvests}{number of harvests.}

\item{nAreas}{number of areas of the plantation site
if not informed the default is 6 areas.}

\item{...}{parameters passed to
\code{\link{testDefVars}} function}

\item{nphases}{number of phenological phases.}

\item{nVars}{number of variables (not counting the dependent variable)
if not informed the default is 3 variables.}
}
\value{
list of areas with harvest data.
}
\description{
Initialized by the \code{\link{testRunDataGen}} function, it generates all
the values of the independent variables and the dependent variable,
returning a list of areas with harvest data.
}
\examples{
\donttest{
nHarvests = 1000
nphases = 5
nAreas = 10
nVars = 6
areas_list  = testBuildSimulationData(nHarvests, nphases)
areas_list_2  = testBuildSimulationData(nHarvests, nPhases, nAreas, nVars)
}
}
\keyword{internal}
