% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{testCreateDataFrames}
\alias{testCreateDataFrames}
\title{Organizes data in dataframes so it can be used in Bayesian learning functions}
\usage{
testCreateDataFrames(data)
}
\arguments{
\item{data}{list of variables referring to an area}
}
\value{
list of dataframes, one for each phenological stage.
}
\description{
Called by the \code{\link{testRunDataGen}} function.
Function receives the list of variables referring to an area and organizes
it into dataframes, one dataframe for each phenological phase.
}
\examples{
\donttest{
area_1_df = testCreateDataFrames(area_1_list)
}
}
