% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{testDefVars}
\alias{testDefVars}
\title{Defines the type and quantity of independent variables}
\usage{
testDefVars(n_var = NULL, type_var = NULL, name_var = NULL)
}
\arguments{
\item{n_var}{number of variables (not counting the dependent variable)
if not informed the default is 3 variables.}

\item{type_var}{a list of types of variables. If informed the \code{name_var}
parameter must be informed too.}

\item{name_var}{a list of names of variables. If informed the \code{type_var}
parameter must be informed too.}
}
\value{
list of variables, that is a list containing
(name of the variable, type of variable, value min and value max)
}
\description{
Called by the \code{\link{testBuildSimulationData}} function.
Function allows defining the type and quantity of initialized independent
variables for data generation. It is possible to define a number of
variables greater than 3. If the type of variable is defined, it is
mandatory to inform a list with the name of the variables as well. If only
the number of variables are defined, the default nomenclature is X_1, X_2,
..., X_n-1, X_n is applied, with randomly defined behavior. If the number of
variables is not defined, 3 variables will be initialized, one that always
grows over time, a constant and one that oscillates over time.
}
\examples{
\donttest{
defProdVariables = testDefVars(...)
defProdVariables = testDefVars(
4,
c("precipitation", "Mn_rate", "Zn_rate", "avg_temp"),
 c(1, 3, 3, 2))
}
}
