% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{testRunDataGen}
\alias{testRunDataGen}
\title{Starts data generation}
\usage{
testRunDataGen(nHarvests, nphases, nAreas, nVars, nClass, ...)
}
\arguments{
\item{nHarvests}{number of harvests.}

\item{nphases}{number of phenological phases.}

\item{nAreas}{number of areas of the plantation site.}

\item{nVars}{number of variables (not counting the dependent variable).}

\item{nClass}{Number of classes for dataframe discretization. Must be 5 or 3.}

\item{...}{parameters passed to
\code{\link{testBuildSimulationData}} function}
}
\value{
list with two dataframes one with continuous data
and the other with discrete data
}
\description{
It receives the parameters for the
\code{\link{testBuildSimulationData}} function
to generate simulated data for carrying out the package tests.
It also distributes the data generated in lists of dataframes representing
the set of phenological phases of each productive area. The data are still
treated so that there is a copy of the data set discretized and another not,
for the purpose of comparing performance between dynamic and static networks.
}
\examples{
\donttest{
areas <- testRunDataGen(nHarvests, nphases, nAreas, nVars, nClass)
}
}
\keyword{internal}
