% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bn_test.R
\name{testRunNetworks}
\alias{testRunNetworks}
\title{TEST FUNCTION - Executes \code{\link{testCreateNetworks}} function to an area}
\usage{
testRunNetworks(arealist, areatype)
}
\arguments{
\item{arealist}{list os dataframes of all phenological phase on
a specific area of the plantation}

\item{areatype}{integer to identify which type of area
the dataframe belongs to}
}
\value{
A dataframe organizing the network metrics generated in the
\code{\link{testRunNetworks}} function. The lines represent the network
performance learned in each phenological phase of the area (\code{arealist})
}
\description{
Executes the \code{\link{testCreateNetworks}} function for all phenological
phases of an area. It also organizes the generated metrics in a dataframe.
}
\examples{
\donttest{
arealist <- list(area1)
metricsArea1 <- testRunNetworks (arealist, 1)}
}
\keyword{internal}
