% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{testSetSimVarValues}
\alias{testSetSimVarValues}
\title{Generates the values of dependent and independent variables}
\usage{
testSetSimVarValues(nHarvests, areatype, prodvars, nPhases)
}
\arguments{
\item{nHarvests}{number of harvests.}

\item{areatype}{integer that defines the type of area. It determines the
kind of relationship between the independent variables and the
'harvest' variable}

\item{prodvars}{list of independent variables, that is a list containing
(name of the variable, type of variable, value min and value max)}

\item{nPhases}{number of phenological phases.}
}
\value{
list of values of all variables
}
\description{
Called by the \code{\link{testBuildSimulationData}} function.
Function starts the  values of dependent and independent variables
The 'harvest' variables is is calculated from rules that depend on the
number of independent variables and the type of area where this
harvest was generated. There are 6 types of areas for standard situations,
3 independent variables, and 6 types of areas for situations where there are
more than 3 variables.
}
\examples{
\donttest{
variables =
testSetSimVarValues(nHarvests, areatype, defProdVariables, nPhases)
}
}
