% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bn_test.R
\name{validateNetworkTest}
\alias{validateNetworkTest}
\title{TEST FUNCTION - Generates Bayesian networks performance evaluation}
\usage{
validateNetworkTest(
  test,
  train,
  dag_fitted1,
  dag_fitted2,
  dag_fitted3,
  dag_fitted4
)
}
\arguments{
\item{test}{dataframe to be used to test the Bayesian networks.
It is composed of a 25% portion of the original dataframe presented to
the \code{\link{createNetworksTest}} function.}

\item{train}{dataframe to be used to train the Bayesian networks.
It is composed of a 75% portion of the original dataframe presented to
the \code{\link{createNetworksTest}} function.}

\item{dag_fitted1}{Fitted Bayesian network to be tested}

\item{dag_fitted2}{Fitted Bayesian network to be tested}

\item{dag_fitted3}{Fitted Bayesian network to be tested}

\item{dag_fitted4}{Fitted Bayesian network to be tested}
}
\value{
List of values returned from \code{bnMetricsMultiVarPrediction}.
See more in \url{https://github.com/KaikeWesleyReis/bnlearn-multivar-prediction-metrics#bnmetricsmultivarprediction}
}
\description{
Using the functions available in the repository
\url{https://github.com/KaikeWesleyReis/bnlearn-multivar-prediction-metrics}
calculates the metrics of the four Bayesian networks generated in
the \code{\link{createNetworksTest}} function when executed in context
of \code{\link{runNetworksTest}} function.
}
\keyword{internal}
